Load Works
Load command works to the best of its abilities

Issues:
Still getting an error on repeat when compiling.  Not a major issue though since the file will still fully compile.  I'm guessing this is just an issue in gulp where it doesn't recognize the typings file.  Could also be that I need to add the typings file when loading the index page